backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pandas==2.2.3
numpy==2.1.3
scipy==1.14.1
yfinance==0.2.54
python-dotenv==1.0.1
openai==1.58.1

